Moving old local plugin manager and registry into plugin module
